refactor: 파일명 변경 및 소켓 연결 분리
